refactor: components and systems to use float for precision in physics calculations

+++ common/ECS/Components/Collider.hpp
@@ -26,8 +26,8 @@ namespace ecs {
         * @brief 2D vector for size and offset.
         */
         struct Vector2 {
-            int x;  ///< X component
-            int y;  ///< Y component
+            float x;  ///< X component
+            float y;  ///< Y component
         };
 
         /**
@@ -40,8 +40,8 @@ namespace ecs {
          * @param mask Bitmask of layers this collider can interact with
          * @param isTrigger True for trigger zones (no physics), false for solid
          */
-        Collider(int sizeX, int sizeY, int offsetX, int offsetY, std::uint32_t layer, std::uint32_t mask,
-                 bool isTrigger)
+        Collider(float sizeX, float sizeY, float offsetX, float offsetY, std::uint32_t layer,
+                 std::uint32_t mask, bool isTrigger)
             : _size{sizeX, sizeY},
               _offset{offsetX, offsetY},
               _layer(layer),
@@ -84,7 +84,7 @@ namespace ecs {
          * @param sizeX New width
          * @param sizeY New height
          */
-        void setSize(int sizeX, int sizeY) {
+        void setSize(float sizeX, float sizeY) {
             _size.x = sizeX;
             _size.y = sizeY;
         }
@@ -94,7 +94,7 @@ namespace ecs {
          * @param offsetX New X offset
          * @param offsetY New Y offset
          */
-        void setOffset(int offsetX, int offsetY) {
+        void setOffset(float offsetX, float offsetY) {
             _offset.x = offsetX;
             _offset.y = offsetY;
         }
